Close pop up window when mouseclick
Set rootClose prop to provide a close mechanism for user in case the
pop up window does not disappear after the mouse cursor leaves the
image. React-bootstrap does not suggest to use 'hover' or 'focus'
triggers for popovers because they have poor accessibility from keyboard
and on mobile devices. Reference:
https://react-bootstrap.github.io/components/popovers/#popovers-trigger-behaviors