Publish & Subscription  - Phase I
Channel, which is defined as system channel or non-system channel. A
publisher can broadcast the messages to all system users when pushing
the message to system channels, or, to the users who subscribed the
non-system channels respectively.

To push the messages to all system users, a specified url is provided
for the system administrator(as a publisher) to do the announcement.

When user(as a publisher) adds [share collections] or [synchronize
collections] act on [Shared with me] and [Synchronized with me], the
system sends the notifications to the subscribers.

User has to acknowledge the notifications, or the system will keep the
notifications in his/her box and remind him/her on next login.

Due to ActionCable is available on Rails 5 and we are now using Rails 4
, and there is no plan to leverage other 3rd party packages. The
current implementation uses client to pull the messages from the server
periodically.