Use correct variable to fix this issue
In function 'create_or_update_attachments', one of the input-types was
changed last time and use the variable name as 'container'. Inside the
function, variable name 'container' was used again when verifying the
update permission, which causes the incorrect update later when updating
attachment's container. It results the unsuccessful dropping files
from inbox.

As a result, define a proper variable name when verifying the attachment
update permission to fix the issue.