Send system upgrade notification to user
Compare the users' local version and the server's version. If a
difference exists, pop up a notification to remind users to reload
the page. This is as a reminder so we do not keep this notification in
the database.

For rspec, check if tag name 'version' exists or not.

A small update on MessagesFetcher.js because of the suggestion from
ESLint(Block must not be padded by blank lines).

Change function 'handleNotification' as a const because of the suggestion from
ESLint.