Deletion of reaction analyses is not working
Before user press Save, user can use Remove icon to set the analysis
is deleted(deletion flag as true) or use Undo icon to bring the
analysis back(deletion flag as false). The analysis is updated to the
database only when user press Save button.

Correct the input(container) of the function handleOnClickRemove.
Use forceUpdate in the function handleUndo because it does not trigger
render properly.