Drop files to Inbox
Add an upload button next to Unsorted folder in Inbox. User can upload
files by using this button and the records are saved in Attachments,
no Containers.

A job is created right after user uploaded files, in order to transfer
the storage from 'tmp' to 'local'. The job is named after user_id.