Research Plan Attachments
Add one new section named 'Files' on UI for attaching files. User can
attach files(from local) or delete/un-attach files by pressing save
button. User can undo his/her deletion before saving.

User can attach files to a Research Plan only when he or she has the
'update' permission of that Research Plan.

If the configuration of OfficeEditor is on, it means the function to
edit the document is provided. A 'Edit' button will be enabled for user
only when the attachment already existed in the DB. (Note: This
configuration will be completed in the OfficeEditor task, not in this
task. The edit button is disabled for now.)

A function to expand or collapse the image is provided. The image is
expanded as default if the image exists, otherwise, it is collapsed.

The information header part is adjusted for better view(not too many
borders).

Take out 'Max size is 50MB' from the upload file message.

Update some syntax based on ESLint.