Split Wellplate
A new wellplate is created with the same attribute-values as selected
wellplate. The new records in table(Well) are also be created to keep
the relationship between new wellplate and 'New Sample'. 'New Sample'
means the new sample created from the original sample on the original
wellplate.

The name of wellplate is required and is ended with surfix '-Split'.