Copy reaction should allow loading input
When creating a reaction with polymer as starting material, the loading
value is always 0 no matter user inputs the value or not. That's the
current logic about reaction creation and will not change.

When copying a reaction with polymer as starting material, the current
logic about the loading value is to set it as null and does not bring
the value from the source one. With this patch, the loading value is
copied from the source reaction on front-end.

There is a parameter 'keepResidueInfo'. No change is applied for the case
(keepResidueInfo is ture) because a code comment said
'// only reset loading'. So, leave it as well.