Analysis preview
User can view the full image by clicking on the 'preview' image(thumbnail)
if the 'preview' image is neither 'Preview not available' nor 'no attachment'.
The cursor of 'preview' image is 'default' if it is 'Preview not available' or 'no attachment'.
The cursor of 'preview' image is 'pointer' if neither 'Preview not available' nor 'no attachment'.

A new component named 'ImageModal' is created to support this function.
The props are:
- hasPop: 'true' to have pop up feature, 'false' no pop(only preview)
- preivewObject: preview object
--- src: display source
- popObject: pop object
--- title: title of pop up window
--- src: display source
--- fetchNeeded: 'true' to fetch the image by fetchId
--- fetchId: image id

The component 'SvgWithPopover' is refactored.
The props are:
- hasPop: 'true' to have pop up feature, 'false' no pop(only preview)
- preivewObject: preview object
--- txtOnly: '' means the image file as preview content, otherwise, text as preview content
--- isSVG: image file is svg type or not
--- src: display source
- popObject: pop object
--- title: title of pop up window
--- src: display source
--- height: height of display
--- width: width of display

An API is created to fetch the image by attachment id.