Analysis png image preview
In database, there are some attachments with 'content_type' as nil so
the preview function can not tell if the file type is image and therefore
it does not process the fetching image file. The preview shows thumbnail
file in the modal.

Use MimeMagic to detect the mime type of a file when container_serialzer
generate json_tree. If the attachment.content_type is nil, try MimeMagic
instead.