Reaction scheme: reference sample to not change when another is deleted
The selected reference sample of Reaction Scheme is changed automatically
after user deletes the other reference sample(not selected one). So now
user needs to review the reference data if he/she just removes some
reference sample which is not the selected reference one.
E.g.
User_A picks one of the samples from 'starting material' as a selected reference.
If User_A removes the sample from 'reactants', the system changes the
selected reference sample to the first one in reactants' automatically.

Update the logic of ReactionDetailsSchema.deleteMaterial. The
updatedReactionForReferenceChange is triggered only when selected reference is
changed. I.e. only consider the deletion from 'starting_materials' and 'reactants'.
The variable reaction.referenceMaterial will not exist if no reference selected
or the reference one is the deleted one. No action will be taken if the
deleted one is not marked as reference.