Net OpenTimeout
Return as nil when exception happen and log the cid and exception for
further investigation.

Note: We have set up the timeout within HTTParty thus we just add exception handling

About Net::OpenTimeout
(refer to https://www.exceptionalcreatures.com/bestiary/Net/OpenTimeout.html)
A successful HTTP request has 3 main steps:

Open a TCP connection to the endpoint.
Send a request over the connection.
Read the response written to the connection.
Net::OpenTimeout is raised when step 1 doesn't succeed within the given time.

This error is not the same as Net::ReadTimeout error which is raised when a connection
is successfully made, but the response cannot be read within a given amount of time
(i.e. when step 3 does not succeed within a given time).

About HTTParty timeout
(refer to https://github.com/janko/the-ultimate-guide-to-timeouts-in-ruby)
HTTParty timeout raises Net::OpenTimeout on connect timeout and Net::ReadTimeout on read timeout