Related changes based on ols file
1. Fix Reaction Type Recently-selected save failure:
a typo (data[name.to_sym] = [term], not [term_id]) caused an exception when saving Profile.

2. Fix 'null' data in database:
for owl_name = 'chmo', the data of synonyms was saved as string 'null'.

3. Fix incomplete content of Reaction Type Recently-selected:
for the one which has no synonym, the content was incomplete.

4. Remove parentheses if there is no synonym/synonyms.

5. Add space into ols_value.value of Profile to solve the warning about duplicated-value; trim data before saving into database.

Note:
Before runing db:migrate, the ols files must exist.