Admin to test datacollector connection
This is part of datacollector upgrade, refer to

Add one new button 'test connection' on Admin UI and
it is only available for sftp method. By clicking 'test connection',
the system connects to host via sftp and returns a message to user to show
connection is built successfully or not.

Additional:
Include notification on AdminHome.
Import startsWith, endsWith from lodash only instead od import the whole lodash (already done in #805).