Admin validate local collector path via whitelist
Add local collector path whitelist in datacollectors.yml:
  :localcollectors:
    - :path: '/home/ftpuser'
    - :path: '/home/claire/tmp'

When user choose 'local' method, i.e. 'filewatcherlocal' or 'folderwatcherlocal',
the whitelist will show.

There is a validation on editing config. function and only apply for method: local.

E.g.

The whitelist is as below:
  :localcollectors:
    - :path: '/home/ftpuser'
    - :path: '/home/claire/tmp'

Set the Dir as '/home/ftpuser', it is validation.
Set the Dir as '/home/ftpuser/xyz', it is validation.
Set the Dir as '/home/claire/../ftpuser', it is validation and will reset to its real path: '/home/ftpuser'.

Set the Dir as '/home/xyz', it is not valid.
Set the Dir as '/home', it is not valid.