Inbox show count and fetch data only on inbox expand
Instead of fetching inbox completed information when the component is mounted,
fetch(and show) inbox count number only.

Fetch inbox completed information when user click on 'Refresh'.
Note: After the component is mounted, the system will fetch
inbox completed information too if it is the first time that user
clicks on 'Inbox'.