Sample or Reaction images are missing after Import Collections
Issue: Reaction svg is missing always
Solved:
(1) Use Zip::File instead of Zip::InputStream. With Zip::InputStream,
the file is not read completely.
(2) Overwrite reaction_svg_file at the end of the method. PS. There is
also a comment about this (# overwrite with the image from the import,
this needs to be at the end...) but do not know why the codes are not
moved as comment.

Issue: Sample svg is missing sometimes
Solved: This issue happens when samples have the same svg file. Keep the
final svg file information to solve this problem.