Wrong field name when import samples
Issue:
User exports samples in excel file format via 'Export samples from selection'
and then he/she tries to import samples with this file via 'Import samples to collection'
but gets data of name and external label are missing.
(Note: user does not make any change on the column name of the file content)

Caused by:
Current 'Import samples to collection' function takes the column name of the file content as
the column_name of database table so the data is missing in some cases.