Reaction analyses abnotmal saving
Issues:
(1) When users make changes on analyses, either of reaction or of sample,
the UI remains 'pending to save' and users are not able to continue
their work on this reaction unless they switch to another element or
close and re-open this reaction.
(2) When users make changes on analyses of reaction, the UI should be
'pending to save' (same as changing analyses of sample).

Fixed by changing updated_at with seconds and calling this.props.parent.handleReactionChange.