Start tidying up the Gemfile
Update bundle to rails 5.0

We commented out the ketcherails gem, since this rails engine does not
support rails 5.0 yet. This engine has to be updated accordingly.

We also commented out the mailcatcher gem because it does not support
rack 2.0 yet. This means that mailcatcher has to be installed outside
the bundle. This should not be a big problem, since mailcatcher runs in
its own process. We will discuss adding a Procfile to the project with
the other team.

(pi_r: rebase  disable nmr_sim)

Update thor gem

This gem was pinned to an old version. We updated this so we would have
the merge option in the app:update task.

We don't really know why thor is in the production group and a project
wide search did not really tell us more. We hope this major version jump
won't be this big of a problem.