Update bundle to rails 5.0
We commented out the ketcherails gem, since this rails engine does not
support rails 5.0 yet. This engine has to be updated accordingly.

We also commented out the mailcatcher gem because it does not support
rack 2.0 yet. This means that mailcatcher has to be installed outside
the bundle. This should not be a big problem, since mailcatcher runs in
its own process. We will discuss adding a Procfile to the project with
the other team.