Update thor gem
This gem was pinned to an old version. We updated this so we would have
the merge option in the app:update task.

We don't really know why thor is in the production group and a project
wide search did not really tell us more. We hope this major version jump
won't be this big of a problem.