Use atomic update when using last
Rails does not expose its internal @records array starting with rails 5.0. So
operating on .last does not work anymore

Add USE_HEAD env variable to specs

This makes it possible to run the feature specs with the browser in the
foreground (with a graphical ui). This is very useful for debugging.

Fix order dependent spec

Fix order dependent spec

Fix flaky spec