Do not check precompilation for assets
This is a quickfix for the ketcherails engine, since its assets cause
production to throw exceptions.

Instead of using this, the ketherails assets should probably add
themselves to the precompile path, but we want to deliver this as soon
as possible so that testing can continue.

Remove ketcherails hotfix

We found the reason for why we needed this hotfix in the first plalce
and fixed the root cause. The root cause was that the image file was
missing in the precompile path within the ketcherails gem.