npm: rm some unused packages
npm webpack: add util, querystring, process
(was provided by browserify)
>

[webpack-cli] ModuleNotFoundError: Module not found: Error: Can't resolve 'util' in '/home/chemotion/workspace/chemotion_ELN/node_modules/aviator/src'

BREAKING CHANGE: webpack < 5 used to include polyfills for node.js core modules by default.
This is no longer the case. Verify if you need this module and configure a polyfill for it.

If you want to include a polyfill, you need to:
	- add a fallback 'resolve.fallback: { "util": require.resolve("util/") }'
	- install 'util'