Add a change history /w logidze
Store a history which keeps track of every change made to certain
entities and add a "History" tab to the Reaction/Sample detail pages.
Related entities are displayed in the version history. E.g. Reaction ->
ReactionsSample

Save timestamp and author of the version with the changes. A change diff
to the previous version can be displayed.

Tracked entities:

- Attachment

- Container

- ElementalComposition

- Reaction

- ReactionsSample

- Residue

- Sample

Co-authored-by: VadimKeller <v.keller@cortona.de>