Fix missing preview images for analyses
The bug was introduced when refactoring the ContainerEntity.
Originally one instantiation of the container entity was used to display the whole container tree.
I considered this code to be really hard to understand, so I refactored it the way that every container
is serialized invididually.

This does not work for the preview image though, as the image file is attached to the dataset container,
while the frontend assumes it to be present on the parent analysis container.

This PR checks if a given container is of type 'analysis' and if so, fetches all eligible images
from the dataset children.