Remove detail_level method from ElementDetailLevelCalculator
As we discussed, we will only use the simpler logic of nested_detail_levels,
based on the maximum configured level over all accessible collections.
This removes a lot of special cases (that were not required anyway).

Further optimization will be possible when the shared and synced collections
feature gets merged, so we only have one of those left.