Set Reporter::Xlsx::ReactionList spec to pending
The spec tries to parse a Xlsx file generated by Axlsx using Roo.
The currently used Roo gem has a parser implementation that does not properly understand
the valid xml structure Axlsx generates, so the spec fails.

The newer version Roo 2.9.0 appears to have addressed this bug but requires Ruby 2.7 to run.
Until we complete the Rails 6 Upgrade, this spec is pending