Fix gem versions for net-ssh and net-scp
The previous gem upgrade commit regressed the net-scp version from 3.0.0 to 1.2.1
This was caused by the upgrade of net-ssh to 7.x.y which is not yet supported by the current net-scp gem.
Downgrading net-ssh a bit allowed to use the latest version of net-scp.

We should upgrade to net-scp 4.0.0 when it reaches stable status (currently 4.0.0rc1)