Improve Dev setup for M1 macs (#1083)
* Improve Dev setup for M1 macs

M1 macs use an ARM-based processor architecture, which will be picked up by docker generating ARM linux images by default.
So we specifically use the linux/amd64 image to create a non-ARM image to work with.  This way all compilation and package installations
(google-chrome-stable does not have ARM packages for example) run the expected way.

Another issue is that docker for mac uses qemu as backend on M1 macs, which does not support inotify. Because of that,
the EventedFileWatcher used in the development environment does not work and had to be replaced by a FileWatcher without usage
of inotify

* Add database setup to run-ruby-dev.sh