fix: camelcasing attributes for proper display of svgs
due to the scrubber library lowercasing all attribute names some properties
are not rendered in the browser.

In this case beads with a gradient  in molecule, sample and reaction
were not displayed in browser and were also missing after conversion in png in doc report.

(successiv gsub benchmarked as faster than gsub with regex and a dict)