Replace react-stickydiv with react-sticky-el
Turns out that react-stickydiv keeps two copies of all its children in
the DOM. One for the regular state and one for the sticky state.

Having two copies of the same radio button rendered (but hidden) in the
actual DOM messes up selecting by label because IDs are no longer
unique.

So this cuts down the number of elements rendered for the element detail
views by 50%.

https://github.com/robbestad/react-stickydiv/blob/2f00573e36615799024299e233b2623a73e1c14b/index.jsx#L78-L85