Fix additional popover Header breaking rounded border
The first <Popover.Header> is not the first DOM node in the <Popover>.
React bootstrap prepends another element for the arrow, so :first-child
does not work here.

I attempted to use `.popover > *:nth-child(n + 2 of .popover-header)`
but I suppose our css preprocessor does not support the relatively new
nth-of selector, because the selector is mangled in the compiled output.

https://developer.mozilla.org/en-US/docs/Web/CSS/:nth-child#the_of_selector_syntax