Rework tab layout editing to proper drag and drop
This patch introduces the TabLayoutEditor component with a
drag-and-drop interface for reordering and hiding/showing tabs.

The component accepts an ItemComponent via the `getItemComponent` prop
for rendering individual items.

With this patch we no longer need to manage the synthetic 'hidden' Tab
when no tabs are currently hidden. This is now the editor component's
responsibility and using code does not have to manage that.

This also removes the `DragDropItemTypes.LAYOUT` constant in favor of a
unique scope for each instance of the editor, because drag-and-drop is
only supposed to work within a specific visible/hidden list combination
and not globally across editors.

State changes are now propagated via the `onLayoutChange` callback and
not by reaching into the state of the component.