fix(UI): minor css / react improvements and fixes (#2421)
* chore: Narrow and define ElementAllCheckbox props API

Instead of passing a complex `ui` prop with undefined keys, we now pass
only what is necessary to the component.

This shows that the formerly used prop `checked` was not provided.

* fix: Do not size inbox modal based on whether the collection tree is shown

* fix: Fix Inbox size dropdown with size `small`

The `.small-panel` class adds `overflow: hidden`, hiding the dropdown.

* fix: Disable rather than hide prev/next buttons in element pagination

This way the buttons jump around less.

* chore: Separate Home and mydb Navigation components

With the new layout, the two components will look and function
differently.

* chore: Separate managing actions from NavigationModal

* chore: Separate export/import context actions from NavigationModal

* fix: Fix selecting items from sample import

While selecting/deselecting would work in production, there were plenty
of errors in the console because the SelectCellRenderer's rowIndex was
not used/passed correctly. Probably because of an API change in
AgGridReact.

* chore: Use ElementStore for managing sdf upload confirmation preview

...instead of the global NavigationModal

* chore: Remove NavigationModal and corresponding UI state

The previous commits make this component obsolete.

* chore: Let context actions handle their own disabled state

* chore: Remove use of `user.profile.data.customClass`

There does not seem to be a way to modify this value and there is no
personally customizable theming in the new design.

* chore: Move FlowViewerModal UI state management out of App component

* chore: React to state changes in SplitElementButton

* chore: Remove setting unused `klasses` state

* chore: Simplify ElementsTableEntries `classnames` usage

* chore: Remove shouldComponentUpdate from ElementsTableEntries

We don't do any heavy calculation during render and practically check
whether anything has changed so there is little need for checking the
few cases where we might not have to re-render.