Consolidate element drag handle logic and component
This moves the logic for figuring out whether a given element can be
dragged into the ElementDragHandle component and unifies it's style. So
no more `fa-home` for `DragDropItemTypes.GENERALPROCEDURE`. :'(

This fixes the issue where drag handles would not properly update when
the current element changed.

It also fixes the issue where it wasn't possible to drag a sample into
the Sample Task Inbox without setting a current element that could
accept samples.