Move element group collapse state into UIStore
With the state in ElementsTable, there was no way to trigger a re-render
of element lists/groups when the state changed, so the result of
`isGroupCollapsed` was only reflected in child components because we
re-render for some other reason when the `currentElement` changes. This
should probably be investigated.

Because DeviceDescriptionList is build differently and does not
re-render, the changes to `isGroupCollapsed` would not show and
collapsing groups would have no visual effect. This is actually the
expected behaviour.

So in order to have the ElementGroupsRenderer reflect the group state
correctly and be able to collapse/expand all groups from within
`ElementsTable`, the group state was moved to UIStore.