Force update to mitigate the fact that we don't properly update state
The `handleLabelChange`, `addSubcollection` and `deleteCollection`
methods mutate the state. :(

This used to not fail because we updated the state on every click. Now
that we don't anymore, we need to either properly update the state and
create a new tree object or `forceUpdate`. Opting for the hacky quick
fix here.