Fix drag & drop preview drift in collection management
The library we are using assumes all items in the tree have the same
height and does not display the preview properly when the don't.

This made rearranging collections in a long list with many synced
collections very hard and at times practically impossible.

To fix this, sync information is accessed via a button and shown in a in
a modal instead of as a separate line below the collection. This gives
every collection in the tree the same height.