fix: drag-n-drop preview drift in collection management (#2494)
* Use `active` state for only one type of object

The `active` state used to contain actual collections and handle
visually highlighting the 'active' collection on click *and* the object
describing the sync of collections and users.

With this patch, the `active` state only handles the sync object and
highlighting is done via css `:hover`.

* Force update to mitigate the fact that we don't properly update state

The `handleLabelChange`, `addSubcollection` and `deleteCollection`
methods mutate the state. :(

This used to not fail because we updated the state on every click. Now
that we don't anymore, we need to either properly update the state and
create a new tree object or `forceUpdate`. Opting for the hacky quick
fix here.

* Fix drag & drop preview drift in collection management

The library we are using assumes all items in the tree have the same
height and does not display the preview properly when the don't.

This made rearranging collections in a long list with many synced
collections very hard and at times practically impossible.

To fix this, sync information is accessed via a button and shown in a in
a modal instead of as a separate line below the collection. This gives
every collection in the tree the same height.