ci: push2deploy (#2610)
The docker-compose.p2d.yml file uses app container (based on modified eln image) with CONFIG_ROLE=combine. The Dockerfile.p2d adds libraries that are useful for compiling gems and node packages, switches the image to the branch from where the build is initiated*, and installs all gems and packages as per need of the branch.

* this in turn needs the .git folder to be included in context. This is the default behaviour, which means that there is no need for the folder to be placed in the .gitignore file (therefore, the change of its contents in this PR). We then are able to force the file to be in context by giving BUILDKIT_CONTEXT_KEEP_GIT_DIR=1 as a build argument.