refactor: update components exposure in sample_entity.rb (#2815)
* refactor: update components exposure in sample_entity.rb

Modified the exposure of the :components attribute in the SampleEntity to include anonymization options when not displayed in the list. This change enhances data privacy while maintaining the use of the ComponentEntity for component representation.

* refactor: preserve solvent when creating new samples in reactions
Remove :solvent from exclusion list in create_new_sample to ensure
solvent data is preserved when creating new materials via ReactionAPI.


ref: #2706


---------

Co-authored-by: Tasnim Mehzabin <tasnim.mehzabin@live.com>
Co-authored-by: Adrian Herrmann <headri@web.de>